pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "telegram-reader"
version = "0.1.2"
description = "Minimal, read-only Telegram reader CLI and library."
readme = "README.md"
authors = [{ name = "Artem Dudarev" }]
requires-python = ">=3.12"
dependencies = [
  "telethon>=1.34.0",
  "PyYAML>=6.0",
  "platformdirs>=4.0.0",
]
keywords = ["telegram", "cli", "reader", "telethon", "agents"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Topic :: Communications :: Chat",
]

[project.optional-dependencies]
test = ["pytest>=8.3.0", "pre-commit"]

[project.urls]
Repository = "https://github.com/dudarev/telegram-reader"

[project.scripts]
telegram-reader = "telegram_reader.cli:main"
tgr = "telegram_reader.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_reader"]

[tool.hatch.version]
path = "src/telegram_reader/__init__.py"
